requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
args==0.1.0
asyncer==0.0.7
attrs==25.1.0
av==14.0.1
beautifulsoup4==4.13.2
bidict==0.23.1
bitsandbytes==0.45.0
braceexpand==0.1.7
certifi==2024.12.14
chainlit==1.3.2
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
Deprecated==1.2.15
diffusers==0.32.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
fairscale==0.4.13
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
fire==0.4.0
flash_attn==2.7.4.post1
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gdown==5.2.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
hpsv2==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
image-reward==1.5
imageio==2.36.1
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
Lazify==0.4.0
lazy_loader==0.4
literalai==0.0.623
loguru==0.7.3
lpips==0.1.4
MarkupSafe==3.0.2
marshmallow==3.24.0
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.0
omegaconf==2.3.0
open_clip_torch==2.30.0
openai==1.59.3
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==23.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
portalocker==3.1.1
POT==0.9.5
propcache==0.2.1
psutil==6.1.1
pyarrow==19.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.9
python-socketio==5.12.1
pytorch-fid==0.3.0
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
safetensors==0.5.0
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
t2v_metrics==1.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.2
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.28.2
urllib3==2.3.0
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
webdataset==0.2.100
wrapt==1.17.0
wsproto==1.2.0
xformers==0.0.29.post1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
